Add a static heap.
This patch adds a static heap (managed as static data) with an overrideable default size. The heap will return NULL when it runs out of space. The default size is 128M.

Turn on this heap with the cmake flag -DSM_STATIC_ALLOC=ON and adjust the size with -DSM_STATIC_HEAP_BYTES=<num-bytes>.

This implementation is correct, but not yet very fast. Furthermore, it's not recommended to be used in multithreaded scenarios yet - its thread safety has yet to be handled.