[ELF] Add basic ELF loader.
This patch adds a super basic ELF loader. The purpose of the loader is as a learning exercise, and as a way to test an ELF linker that I will work on someday.

All it can do for now is run a simple test executable that returns `4 + argc` - no dynamic linking support or anything has been added at this point. It also ignores useful extensions, and is probably wrong in many ways, but the test program does work on macOS!